Added a few backgrounds to test the background rotation code. The backgrouns are really ugly gradients at the moment, artists are welcome to design something better.
A waterfall, or a large cave, or rocks or dirt would be good candidates; since
the background should not be too distracting, and since it is re-rendered at
window resize time and at level loading time, it's better for it to not be too
complex. The geek-city background does not seem slow to load on my laptop, so
that may be a good reference point.

svn path=/trunk/KDE/kdegames/kgoldrunner/; revision=687287