Split theme graphics in a "Set" (as in cinematographic set) and "Actors" parts.
Added a second background to the geek city theme (it's a copy of the first one,
painted blue instead of green. Someone more artistically inclined may even move
some of the components around...

Tweaked the coin in the default theme and changed the black with a dark blue in
the backgrounds, so that enemies will show.

Maybe it would be useful to further split the actors in hero, villain and
villain with gold, and use identical names for the frames. In this way, it
would be possible to, say, copy the hero to the villain, tweak the template
body parts (change colors, add different details...) and have a decent
animation that could be further tweaked if needed.

CCMAIL: kde-games-devel@kde.org
CCMAIL: Ian Wadham <ianw2@optusnet.com.au>


svn path=/trunk/KDE/kdegames/kgoldrunner/; revision=698126