Add a dig-while-falling option to recording and game-data files.
Allowing the hero to dig bricks while falling past them was a bug in
KGoldrunner. However, some contributors of levels exploited this in the
games they composed, so it became a feature. This patch opens the way
to restrict use of dig-while-falling in specific games and levels, such
as the Tutorial and Initiation games.