Lowercase for the desktop file (and the dbus interface)
It looks like the desktop file should match the name of the executable
(with the domain), also - but not only to allows the task manager to
recognize the program under Wayland.
Not sure about the DBus interface, but the capitalization did not match
anyway the name of the exposed interface.